All needed dlls are in build output, but the application was still unable to load the DatabaseLayer dll, the reason was mybe the missing references in ApplicationLayer.deps.json this file has no reference to DatabaseLayer.dll and npgsql.dll.
for this reason i added a "WORKAROUND CODE" to load this libs by convention from my bin folder.

now the application is working like it should, and like it works on net462!

the question is why this dependencies are not resolved automaticaly on runtime, because they don't even create a warning or error on compiletime